+++ applications/greetings/src/main/java/org/phoebus/applications/greetings/GreetingsApp.java
@@ -3,23 +3,18 @@ package org.phoebus.applications.greetings;
 import org.phoebus.framework.spi.AppDescriptor;
 import org.phoebus.framework.spi.AppInstance;
 
+@SuppressWarnings("nls")
 public class GreetingsApp implements AppDescriptor {
 
     public static final String Name = "Greetings";
 
+    @Override
     public String getName() {
         return Name;
     }
 
-    public void start() {
-    }
-
-    public void stop() {
-    }
-
     @Override
     public AppInstance create() {
-        return null;
+        return new Greet(this);
     }
-
 }
