+++ applications/greetings/src/main/resources/META-INF/services/org.phoebus.framework.spi.ToolbarEntry
@@ -1,2 +1,2 @@
-org.phoebus.applications.greetings.Greet
+org.phoebus.applications.greetings.GreetToolbarEntry
 org.phoebus.applications.greetings.PersonalGreet
